+++ src/components/_classes/component/editForm/Component.edit.logic.js
@@ -369,7 +369,7 @@ export default [
               {
                 weight: 20,
                 input: true,
-                label: 'Schema Defenition',
+                label: 'Schema Definition',
                 key: 'schemaDefinition',
                 editor: 'ace',
                 as: 'javascript',